+++ app/Http/Controllers/FollowerController.php
@@ -29,6 +29,10 @@ class FollowerController extends Controller
      */
     public function create($id)
     {
+        $userJiinId = Auth::guard('web')->user()->jiin_id;
+        if (empty($userJiinId)) {
+            throw new Exception('ログインユーザーの寺院IDが取得できませんでした。');
+        }
         //家族続柄のデータ取得
         $relationships = Code::query()
                             ->where('key1', '=', 'RELATIONSHIP')
@@ -108,6 +112,10 @@ class FollowerController extends Controller
      */
     public function show($id)
     {
+        $userJiinId = Auth::guard('web')->user()->jiin_id;
+        if (empty($userJiinId)) {
+            throw new Exception('ログインユーザーの寺院IDが取得できませんでした。');
+        }
         //寺役職のデータ取得
         $positions = Code::query()
                         ->where('key1', '=', 'POSITION')
@@ -138,6 +146,7 @@ class FollowerController extends Controller
                                  'followers.occupation',
                                  'followers.memo')
                         ->where('followers.id', '=', $id)
+                        ->where('followers.jiin_id', '=', $userJiinId)
                         ->first();
 
         // $follower = Follower::find($id);
@@ -154,6 +163,10 @@ class FollowerController extends Controller
      */
     public function edit($id)
     {
+        $userJiinId = Auth::guard('web')->user()->jiin_id;
+        if (empty($userJiinId)) {
+            throw new Exception('ログインユーザーの寺院IDが取得できませんでした。');
+        }
         //家族続柄のデータ取得
         $relationships = Code::query()
                             ->where('key1', '=', 'RELATIONSHIP')
@@ -209,6 +222,11 @@ class FollowerController extends Controller
             $follower->chiefmourner_flg = 0;
             $follower->deceased_flg = 0;
 
+            if (Auth::guard('web')->check()) {
+                $follower->jiin_id = Auth::guard('web')->user()->jiin_id;
+            } else {
+                throw new Exception('ログインユーザーの寺院IDが取得できませんでした。');
+            }
             $follower->save();
 
             //正常に登録出来たらコミット
@@ -237,6 +255,10 @@ class FollowerController extends Controller
     {
         DB::beginTransaction();
 
+        $userJiinId = Auth::guard('web')->user()->jiin_id;
+        if (empty($userJiinId)) {
+            throw new Exception('ログインユーザーの寺院IDが取得できませんでした。');
+        }
         try{
             Follower::find($id)->delete();
